Add a "Distro" column to the "db jobs" output
The openSoftware developers build packages for multiple distributions
and requested an additional column in the `butido db jobs` output to
show for which distro the job was built (without having to perform a
separate/manual lookup for each entry). This feature should of course be
useful in general. The distro information is based on the name of the
container image. Some of our image names are quite long so we decided to
change the format of the `images` configuration setting (108eecc) so
that users can specify a short name that will be used to shorten the
output/column of the `db jobs` command.
The full/long image name, that is stored in the DB, will be used as a
fallback if there is no corresponding entry in the configuration (e.g.,
because the image was dropped in the meantime).

This resolves #48.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>