clap: Call arg.action(ArgAction::SetTrue) for flags
In a98716e, I noticed that the instructions for switching away from the
deprecated ArgMatches::is_present method are:
> For flags, call arg.action(ArgAction::SetTrue) and lookup the value
> with matches.get_flag(...)

This commit handles that first part. I only touched the flags that are
used via the `get_flag` method and ensured that no usages of the
deprecated `is_present` method are left.
This should cover all flags but some might be left if they're accessed
differently (I'll try to catch those when I do a larger refactoring of
our clap code (I noticed other issues/inconsistencies as well)).

Documentation and examples:
https://docs.rs/clap/3.2.23/clap/builder/enum.ArgAction.html#variant.SetTrue
https://docs.rs/clap/3.2.23/clap/builder/struct.Arg.html#method.new
https://docs.rs/clap/latest/clap/_tutorial/index.html#flags
https://docs.rs/clap/latest/clap/_derive/_tutorial/index.html#flags

Signed-off-by: Michael Weiss <michael.weiss@atos.net>