Change the default log level from info to warn
In ff5666c, we switched from env_logger to tracing. We later noticed
that, as a side effect, the default log level changed from error [0] to
info [1]. The info log level is too verbose by default but warnings
might be useful or even important so we decided to explicitly setting
the log level to warn.
The `EnvFilter` is required so that the `RUST_LOG` environment variable
is still considered [2].

Fix #191.

Co-authored-by: Nico Steinle <nico.steinle@atos.net>
Signed-off-by: Nico Steinle <nico.steinle@atos.net>
Signed-off-by: Michael Weiss <michael.weiss@atos.net>

[0]: https://docs.rs/env_logger/0.10.0/env_logger/#enabling-logging
[1]: https://docs.rs/tracing-subscriber/0.3.16/tracing_subscriber/fmt/struct.Subscriber.html#associatedconstant.DEFAULT_MAX_LEVEL
[2]: https://docs.rs/tracing-subscriber/0.3.16/tracing_subscriber/fmt/struct.SubscriberBuilder.html#method.with_max_level