clap: Stop setting the version of commands (cleanup)
This was only relevant for the help output since we disabled the version
flag for all subcommands in 43aac56 (for the main butido command it is
now implemented manually). However, the help output doesn't include the
version since the update to clap version 4 in 2acf402 so setting the
version shouldn't have any effect anymore.

It was also unnecessary to set the same version for each subcommand as
there's a `propagate_version` method now [0] (so we should use that
instead if we ever need the version again).

Note: The clap documentation still states that the `version` method
"sets the version for the short version (-V) and help messages" [1]
but the second part (help messages) isn't correct anymore.

[0]: https://docs.rs/clap/4.1.13/clap/builder/struct.Command.html#method.propagate_version
[1]: https://docs.rs/clap/4.1.13/clap/builder/struct.Command.html#method.version

Signed-off-by: Michael Weiss <michael.weiss@atos.net>