Drop the version constraints for encoding_rs
This was done in 947f0ce as version 0.8.30 of encoding_rs apparently
broke the `cargo deny check licenses` check due to licensing information
changes.
We did, however, already update to 0.8.31 in b75231e and the cargo-deny
licensing check seems to be fine now.
It looks like it was indeed fixed again in version 0.8.31 [0] (all
changes: [1]).

[0]: https://github.com/hsivonen/encoding_rs/commit/fcf1e106d3402ee506e0c5419a83d5d25bd650c9
[1]: https://github.com/hsivonen/encoding_rs/compare/v0.8.30...v0.8.31

Signed-off-by: Michael Weiss <michael.weiss@atos.net>