Drop the dependency on funty
The direct dependency on funty was added to Cargo.toml in db698a4 to
prevent a build failure as the `bitvec` crate depended on `funty` but
wasn't compatible with a newer version than 1.1.0 due to breaking
changes in `funty`.
But we currently need neither `funty` nor `bitvec` (apparently both of
them were transitive dependencies but I cannot trace their origin as
they've been dropped since before the Cargo.lock file was added to the
Git repository in 2506305).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>