Disable all of ptree's default features
The ptree crate is pretty unmaintained [0] (last commit at the end of
2021 and no response to an open issue) and not really widely used [1].
We'll probably get rid of it eventually (we already have a draft) but
for now I want to avoid the dependency on "atty" which is unmaintained
and has a low severity GHSA for Windows (GHSA-g98v-hv3f-hcfr) that
resulted in a dependabot alert [2].

The "atty" crate is only required for the "ansi" feature but we actually
need none of the three default features:
- "ansi": For advanced text formatting
- "conf": For user configuration (loaded from a user configuration file)
- "value": Implementation of TreeItem for serde_value::Value

Disabling all three features saves us from even more unnecessary and old
dependencies.
No code changes are required and the `butido tree-of $pkg` output
remains the same.

[0]: https://gitlab.com/Noughmad/ptree
[1]: https://crates.io/crates/ptree
[2]: https://github.com/science-computing/butido/security/dependabot/9

Signed-off-by: Michael Weiss <michael.weiss@atos.net>