Resolve all warnings detected by clippy 0.1.73 (beta toolchain)
This allows the `redundant_locals` lint [0] for a function where it's
usage is convenient for moving only a single variable into a closure
while still borrowing the captured variables from the environment (there
are alternatives but they're more verbose).

The other lint [1] detected a manual implementation of both `PartialOrd`
and `Ord` for the `Package` struct. For correctness, the implementation
of `Ord` should be reused for `PartialOrd` to ensure that they always
behave the same.

[0]: https://rust-lang.github.io/rust-clippy/master/index.html#/redundant_locals
[1]: https://rust-lang.github.io/rust-clippy/master/index.html#/incorrect_partial_ord_impl_on_ord_type

Signed-off-by: Michael Weiss <michael.weiss@atos.net>