Deny unknown fields in package configurations
This helps catching errors in package configuration files (`pkg.toml`)
by rejecting unknown fields when loading the packages repository.
The `config` crate is a bit limited in this regard but luckily this can
be caught during deserialization via Serde [0][1].

The error output contains the path of the problematic package
configuration file, the name of the unknown field, and a list of the
valid/expected fields.

Fixes #188.

[0]: https://github.com/mehcode/config-rs/issues/450
[1]: https://serde.rs/container-attrs.html#deny_unknown_fields

Signed-off-by: Michael Weiss <michael.weiss@atos.net>