Remove the `docker.verify_images_present` configuration setting
This setting is too useless for the following reasons:
- It only applied to the build command (but there are additional
  commands that support the `-I`/`--image` parameter)
- The second part that is documented in the example configuration isn't
  implemented (so the setting is misleading)
- The setting should usually be set to `true` (it is only supported for
  the build command and I don't see a good reason why it should be
  disabled for that)
- The implementation is now redundant/unnecessary as the new
  `resolve_image_name` function always performs the first check

We could consider keeping the setting and fixing the implementation
(i.e., use it for `resolve_image_name` and implement the endpoint check)
but it might be better to drop it to make things simpler as we didn't
miss the functionality so far. And IMO the setting isn't really that
useful for disabling the `docker.images` check as one could also
(temporarily) change the `docker.images` setting instead of
`docker.verify_images_present`.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>