Improve the error message when parsing package dependency specifications
The old error message lacked a lot of context so it was really hard to
find the problematic specification in the package repo source code.
This makes it quite a bit easier but it's still far from perfect
(ideally, we'd emit the error while parsing the files in the packages
repository so that we can output the file path and line number as well).

Before this change:
```
Error: tree-of command failed

Caused by:
    0: A package version constraint must have a comparator and a version string, like so: =0.1.0
    1: Failed to parse package version constraint
    2: Mismatch at 0: expect: 61, found: 60
```

With this change (one can `grep` for "libXYZ <1" in the packages repo):
```
Error: tree-of command failed

Caused by:
    0: Could not parse the following package dependency string: libXYZ <1
    1: A package version constraint must have a comparator (only `=` is currently supported) and a version string, like so: =0.1.0
    2: Failed to parse the following package version constraint: <1
    3: Mismatch at 0: expect: 61, found: 60
```
(note: 61 is the decimal representation of `=` and 60 of `<`)

Signed-off-by: Michael Weiss <michael.weiss@atos.net>