Add the dependency type to the package DAG (stored via edge weights)
This will be required for a change in the `tree-of` command to
differentiate between build and runtime dependencies in the output.
For this PoC, I kept the `i8` type for the edge weights (the edge
weights were always 0 / unused before) but we should of course use the
proper `DependencyType` enum type.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>