Fix setting the default log level to WARN
In 76150f0, we used `with_max_level()` to set the default log level but
apparently the behaviour changed in the meantime (or we didn't test it
correctly back then). The documentation of `with_env_filter` [0] states:
> If a filter was previously set, or a maximum level was set by the
with_max_level method, that value is replaced by the new filter.

The idea in 76150f0 was that `with_env_filter` would only override the
default from `with_max_level` if the `$RUST_LOG` environment variable is
set but the `EnvFilter::from_default_env()` also had a default log level
(unfortunately it defaults to ERROR instead of WARN).
We now replace `from_default_env()` with the underlying code [1] but
changed the default log level from ERROR to WARN.

[0]: https://docs.rs/tracing-subscriber/0.3.17/tracing_subscriber/fmt/struct.SubscriberBuilder.html#method.with_env_filter
[1]: https://docs.rs/tracing-subscriber/0.3.17/tracing_subscriber/filter/struct.EnvFilter.html#method.from_default_env

Signed-off-by: Michael Weiss <michael.weiss@atos.net>
Co-authored-by: Nico Steinle <nico.steinle@atos.net>