Add a clippy allow exception to make CI happy/green
In bebe720, I already mentioned that there are two clippy warnings left
and that a fix got merged - I just failed to realize that the fix
probably won't be backported to the toolchain version 1.75 (given that
the fix isn't that important as the lint only emits warnings by default
and the flagged usage is probably not that common).
Anyway, this makes the CI checks green again (they're currently red as
an optional test fails - this test will also become required once
version 1.75 hits the stable channel (unless the clippy fix would be
backported before that)).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>