Add a "release list" subcommand (alias for "db releases")
The "butido release" command offers the subcommands "new" and "rm" but a
subcommand to list all releases was missing as that functionality is
available via the "butido db releases" command.
This isn't really intuitive so let's simply add "release list" as an
"alias" of the "db releases" command by reusing the existing code.

This resolves #143.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>