Refactor the code to list artifacts from the DB
The functionality should remain the same (there are two very minor
differences: the error handling has changed slightly and the usage of
`into_boxed()` incurs a minor performance penalty [0] but other commands
already use it as well).

This refactoring is only useful to extend the command later for
additional filtering options (plus it should be more readable).

[0]: https://docs.diesel.rs/2.0.x/diesel/prelude/trait.QueryDsl.html#method.into_boxed

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>