Stop using result-inspect from Rust version 1.76 onwards
This is required to fix "cargo check" with Rust version 1.76+ (currently
in the beta channel) while still supporting older toolchain versions.

The `std::result::Result::inspect` method used to be a nightly-only
experimental API [0] but got stabilized in Rust version 1.76 [1] so we
don't need to depend on the `result-inspect` crate anymore (we could
make it optional but let's just wait until we bump the MSRV accordingly
and drop it entirely).

[0]: https://doc.rust-lang.org/1.74.0/std/result/enum.Result.html#method.inspect
[1]: https://doc.rust-lang.org/beta/std/result/enum.Result.html#method.inspect

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>