actions: Only pin the checkout action to a major version
The `actions/checkout` action is from a trusted source (official GitHub
Actions), this is in line with the documentation/examples [0], and it
enables us to automatically get useful updates [1]:
> If the action publishes major version tags, you should expect to
receive critical fixes and security patches while still retaining
compatibility. Note that this behavior is at the discretion of the
action's author.

This especially makes sense since we already use this approach for all
other included actions [2] (so it's finally consistent as well!).
(Note: For 3rd-party actions it could make sense to use a more secure
approach (commit SHA) but we don't need any secrets for CI and the
sources should be trustable so our current approach should be fine.)

[0]: https://github.com/actions/checkout/blob/v4/README.md
[1]: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idstepsuses
[2]: See `git grep -hoE '[^ ]+@[^ ]+' -- .github/workflows/ | sort -u`

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>