Make the `compatibility` setting independent of the butido version
Ideally, the butido version should be bumped when there are breaking
changes but this new approach offers the following advantages:
- The compatibility setting doesn't need to be changed when we release a
  new butido version (e.g., with breaking changes that are unrelated to
  the configuration).
- An independent configuration version is more meaningful: We can update
  it immediately when we add a breaking change (no need to make a new
  release), inform the user of the necessary changes, and don't "have"
  to bundle breaking changes.
- One minor issue was that, e.g., configuration version 0.4.1 wouldn't
  be compatible with butido 0.4.0 (but that can be avoided by always
  setting `compatibility` to `x.y.0`, which is compatible to butido
  version `x.y.*`).
- The new type is stricter (`semver::VersionReq` is a SemVer version
  requirement and can be "*", ">=1.2.3, <1.8", etc.).

This fixes #187.

PS: We should also start collecting "upcoming" breaking/relevant changes
in the changelog so that we don't have to go through the commit history
(all of the commits since the last tag) when preparing a new release.

PPS: I chose to start with a `compatibility` setting of `1` so that we
could use `0` for a "smoother transition", if necessary.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>