Deny unknown fields in butido's configuration
This helps catching errors/typos in butido's configuration as well as to
cleanup settings that don't exist anymore or got renamed.
It already revealed a misnamed key in our example configuration file
(that caused the corresponding test to fail)!

We already implemented this for package definitions in b075e97.

The error output contains the name of the unknown field and a list of
the valid/expected fields.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>