Add a helper for checking if the configured directories exist
I already made a mistake (accidentally introduced a `!` - luckily it was
caught by CI for other reasons) in one of the three checks when
introducing `skip_filesystem_checks`.
This simplifies the code and ensures that all checks behave the same and
I already noticed another mistake that I'll fix next (not part of this
commit to keep it a refactoring).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>