Fix the comments for the `FileSystemRepresentation` tests
These comments represent the file tree of the used test repos for the
tests and are only there for better readability. Unfortunately the
comments didn't match the `FileSystemRepresentation` structures for the
tests that they represent (the `baz` directory is always a subdirectory
of `bar` and some `pkg.toml` files were missing). This should be fixed
now (assuming I didn't make a mistake). I also included the used example
content of the `pkg.toml` files.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>