Add a `FileSystemRepresentation` test for loading the example repo
This also fixes loading a packages repository from a path that isn't the
current directory (this failed due to using paths relative to the repo
root [0] - we never noticed this issue since butido loads repos from the
current working directory).
The usage of relative paths seems problematic or rather dangerous but at
least `FileSystemRepresentation` should be fine now (we might switch to
absolute paths in the future).

The test is useful to verify loading a more complex repo and I'll add
another test to verify the example repo (which prompted me to add this
lower-level test as well).

[0]: The error message is also somewhat misleading due to the usage of
relative paths (relative to the repo root):
```
Error: Reading file from filesystem: a/pkg.toml

Caused by:
    No such file or directory (os error 2)
```

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>