Stop stripping the repo root prefix for `FileSystemRepresentation` files
Stripping the repo root prefix from the file paths would be fine if they
remained internal/private to the `FileSystemRepresentation`
implementation, where we could handle everything accordingly, but
they're exposed and it's unrealistic (too error prone) to ensure those
modified file paths will get handled correctly everywhere else in the
code.
Using paths relative to the repo root looks nice but let's favor
correctness. I noticed this when trying to load the example repo in
`examples/packages/repo/` as this would result in various errors, e.g.,
the one from de4c602, and this one:
```
--- repository::repository::tests::test_load_example_pkg_repo stdout ---
Error: Patch does not exist: s/./foo.patch
```

This didn't became a problem before because butido currently only loads
the repo from the current working directory (".") but this is an
undocumented limitation and can easily pose problems in the future
(e.g., now that I'm about to add a test that loads the example repo and
when we'd switch to absolute paths, which could help with the error
messages since only absolute paths make it clear where butido is looking
for the files).

A possible alternative could be to use dedicated types for the relative
paths that keep track of the repo root prefix and automatically handle
or force using the necessary conversions.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>