Add a test for loading the included example packages repository
We should verify that the example repo is and remains valid and this
also serves as a useful test with a more complex repo.

This should resolve #23 (we already have such an example packages repo
in this butido Git repo and this test does at least verify that it can
be loaded, i.e., that the definitions are syntactically valid (regarding
the type checking) - ideally we'd also have a high-level test that
actually runs some builds to verify the logic (semantics) but that can
be done as part of #109 / the CLI tests in #43).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>