Update all dependencies (Cargo.lock) and bump the MSRV to 1.74.0
This is simply the result of running `cargo update` without touching
Cargo.toml (required since we currently only use dependabot to update
direct dependencies and not indirect/transitive dependencies and useful
to avoid a lot of dependabot PRs for minor/patch updates that can be
bundled, like in this commit).

Version 0.7.0 of the `clap_lex` crate requires rustc 1.74 or newer so
let's also bump the MSRV from 1.70.0 to 1.74.0.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>