Fix the "merging" of `patches` arrays in package configurations
The old logic silently ignored a patch if a patch with the same name was
declared in an upper level/layer package configuration.
This fixes that issue and I now know why that "else if" branch exists:
The `config.set_once("patches", config::Value::from(patches))` didn't
last beyond the next `config.merge()` so the custom logic is required to
ensure that `patches` will be empty so that `patches_before_merge` (with
the value from `config.set_once()`) gets used instead.

I'm now using `config.merge()` to properly update `patches` so that we
can simply check if `patches_before_merge == patches` (i.e., no changes
occurred after merging the next `pkg.toml` file).
This looks a bit more complex at first but can be cleaned up next and I
hope that newer versions of the `config` crate make this easier for us
(we can really simplify the code if we can request the source (file path
of the `pkg.toml` file) that set the final `patches` value).

I can also "fix" the tests now that the "merging" works as intended.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>