Update the `config` crate from 0.11.0 to 0.12.0
This is long overdue... The API has changed significantly with the
introduction of the new `ConfigBuilder` abstraction.
Most of the changes were still quite straightforward but the "merging"
of the package configuration "patches" is far from optimal now - the
current approach might unfortunately be quite inefficient (shouldn't
really be that noticeable though). At least this only affects the
performance and not the correctness. I'll try to replace the current
approach with a clean one when updating to version 0.14.0 of the
`config` crate (then we should be able to use `origin` instead and avoid
that whole `fold` logic for "merging" the patches).

Note: The error and additional error context isn't required for
`config_builder.add_source()` anymore as the error from
`config_builder.build()` will contain the file path, e.g.: "invalid TOML
value, did you mean to use a quoted string? at line 1 column 9 in
../.config/butido/config.toml".

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>