Cleanup the logic for "merging" package "patches" (config setting)
The old logic was pretty complex and only necessary to build the correct
relative paths to the patch files (the paths in `pkg.toml` must be
prepended with the relative path to the directory containing the
`pkg.toml` file).
Since the recently released version 0.14.0 of the `config` crate we can
access the "origin" of a configuration value (`Value::origin()`) so we
can use that information to avoid having to check if the `patches` have
changed every time we merge another `pkg.toml` file.

Unfortunately this does currently require a dedicated source
implementation (`PkgTomlSource` but actually why not) since
`config::File::from_str()` always sets the URI/origin to `None` and the
new `set_patches_base_dir()` function is a bit of a hack...
IMO the new code is much more readable, more efficient, and overall
still cleaner though (most of the new code is for error handling and the
custom `Source` implementation).

[0]: https://github.com/mehcode/config-rs/blob/0.14.0/CHANGELOG.md#0140---2024-02-01

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>