Allow dependency names that start with a number
It was possible to declare and build packages where the package name
starts with a digit but it wasn't possible to declare them as
dependencies of another package.
This fixes #356 and I'm also adding a test for it to avoid regressions.

Now we could package `0ad`, `389-ds`, `6tunnel`, `7z`, `9pfs`, etc. \o/

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>