Simplify the dependency parsing regex (unnecessary escaping)
The escaping shouldn't be required for "." and "*" inside a bracketed
character class (`[...]`) and we can avoid it for "-" if we place it at
the beginning or the end of the bracketed character class. The escaping
was also a bit confusing as "\" had to be escaped with another backslash
as it's a Rust string (raw strings would be better in this case).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>