Avoid unnecessary warnings with older Clippy versions
In 80e0101, I first added a rule for a new Clippy lint (introduced in
1.73.0) to fix our optional CI check that lints with the beta toolchain.
This would break the required CI check against the MSRV so I decided to
allow unknown lints for that check in ef29d94. However, this isn't ideal
as it only applies to the CI checks and still shows the warnings when
developing or rather linting locally.
Luckily we can use the rustversion crate to make the Clippy lint
overrides via attributes also depend on the Rust/toolchain version (this
especially helps with cases like in 0556536 where the override is only
required for a single toolchain version).
It's just a bit unfortunate that I cannot use `stable(1.77)` since that
really only applies to stable versions and `1.77` is currently still a
beta version (so I have to combine `all`, `since`, and `before` to
produce a `version(1.77)`).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>