Add tracing instrumentation to Orchestrator
The tracing "instrumentation" (yes, in quotes) that was used before this
patch was not really an instrumentation. It was rather a better form of
using the `log` crate.

This patch introduces some spans and associated tracing events for the
"meat" of the orchestrator algorithm.
Also, the tracing events now have fields instead of interpolating some
values in the trace message.

The tracing here is not really soffisticated yet, but it should be miles
better than the tracing before.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>