Bump the MSRV to 1.76.0 to fix Clippy warnings
We don't have to bump the MSRV yet but this is the cleanest solution to
avoid hacks for the new `incompatible_msrv` Clippy lint [0] that got
added in 1.78.0. We'll probably have to bump the MSRV soon anyway (for
dependency updates) so let's just do it right away to get rid of those
false positives (we use `result_inspect::ResultInspect` for older Rust
versions but Clippy can't consider the `#[rustversion::since(1.76)]`).

[0]: https://rust-lang.github.io/rust-clippy/master/index.html#/incompatible_msrv

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>