Ensure that relative paths for patches are used internally
The rest of the code that processes custom package patches (defined in
`pkg.toml`) only works correctly when using relative paths.

I noticed this issue [0] when attempting to build the packages in
`examples/packages/repo/` which resulted in the following build error:
```
$ butido db log-of 6116589e-471d-4b80-b446-1fd028e82623
Does not exist: /patches//tmp/butido-test-repo/s/19.0/foo.patch
```
The tracing output reveals the issue: "Copying patch
/tmp/butido-test-repo/s/19.0/foo.patch to container at
/tmp/butido-test-repo/s/19.0/foo.patch"
The second path should be prefixed with "/patches/" but that didn't work
because `path1.join(path2)` results in just `path2` if the latter is an
absolute path.
I decided to simply switch to relative paths for now as the path to the
`pkg.toml` repo shouldn't get "leaked" into the containers anyway, at
least IMO. Those relative paths should be good enough for now and still
guarantee that there won't be name collisions (vs. when only using the
file name of the patch).

[0]: This issue must be a regression but I'm not sure how old the
regression is as a quick look didn't reveal it's source and it could
even come from dependency updates (the absolute paths come from `git2`).
One could use git-bisect but I didn't bother yet.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>