Output the IDs of endpoints and container images (endpoint subcommand)
This is basically a quickfix for our `cargo check` CI test against the
beta toolchain that reported that those two fields don't need `pub` as
they're never accessed from outside the module.
Alternatives would be to drop `pub` or add an `#[allow(dead_code)]`
exception but outputting the IDs might actually be for the best (they're
normally not that relevant but could be useful, e.g., when interacting
with the DB).
The main problem is that the `butido endpoint containers list` output
becomes too wide and the IDs get truncated but ideally we'd try to
shorten the IDs anyway (the only problem is ensuring that they remain
locally unique...).

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>