Remove an unused configuration setting
It's unlikely that we'll need it anytime soon and we can always re-add
it if so - there shouldn't be any unused settings around.

Theoretically this is a breaking change that warrants incrementing
`CONFIGURATION_VERSION` but it's unlikely that anyone uses this setting
given that it wasn't even in the example configuration files and the
generic error message for unknown settings should be fine anyway.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>