Add a configuration setting for limiting the DB queries (default limit)
It is very useful to have a default limit for DB queries that can return
lots of results/matches. Without a limit, the listing of DB tables can
take a long time and even result in OOM kills (the latter should ideally
be avoidable via data streaming instead of allocating memory to store
all results before outputting them, which would also help with the
former - at least with the initial delay before the output starts).
See also [0].

This adds a configuration setting for the default limit as it makes
sense to support user/project specific default limits.

Signed-off-by: Michael Weiss <michael.weiss@eviden.com>

[0]: https://github.com/science-computing/butido/pull/323#issuecomment-1836754840